--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -94,7 +94,6 @@ jobs:
         with:
           path: ~/.npm
           key: ${{ runner.os }}-node-${{ hashFiles('asssets/package-lock.json') }}
-      - run: mkdir -p tmp/handlebars
       - run: npm ci --prefix assets
       - run: npm run build --prefix assets
       - name: Push updated assets