--- a/assets/build/build.js
+++ b/assets/build/build.js
@@ -2,125 +2,88 @@ const path = require('node:path')
 const process = require('node:process')
 const cp = require('node:child_process')
 const esbuild = require('esbuild')
-const util = require('./utilities')
-const loadHandlebars = require('./load-handlebars')
+const fsExtra = require('fs-extra')
+const fs = require('node:fs/promises')
+const handlebars = require('handlebars')
+const util = require('node:util')
+
+const exec = util.promisify(cp.exec)
 
 const watchMode = Boolean(process.env.npm_config_watch)
 
-/**
- * Configuration variables
- */
-
-// Basic build configuration and values
-const commonOptions = {
-  entryNames: '[name]-[hash]',
-  bundle: true,
-  minify: true,
-  logLevel: watchMode ? 'warning' : 'info'
-}
-const epubOutDir = path.resolve('../formatters/epub/dist')
-const htmlOutDir = path.resolve('../formatters/html/dist')
-
-/**
- * Build: Plugins
- */
-
-// Empty outdir directories before both normal and watch-mode builds
-const epubOnStartPlugin = {
-  name: 'epubOnStart',
-  setup (build) { build.onStart(() => util.ensureEmptyDirsExistSync([epubOutDir])) }
-}
-const htmlOnStartPlugin = {
-  name: 'htmlOnStart',
-  setup (build) { build.onStart(() => util.ensureEmptyDirsExistSync([htmlOutDir])) }
-}
-
-/**
- * Build
- */
-
-// ePub: esbuild options
-const epubBuildOptions = {
-  ...commonOptions,
-  outdir: epubOutDir,
-  plugins: [epubOnStartPlugin, loadHandlebars],
-  entryPoints: [
-    'js/entry/epub.js',
-    'css/entry/epub-elixir.css',
-    'css/entry/epub-erlang.css'
-  ]
-}
-
-// ePub: esbuild (conditionally configuring watch mode and rebuilding of docs)
-if (!watchMode) {
-  esbuild.build(epubBuildOptions).catch(() => process.exit(1))
-} else {
-  esbuild.build({
-    ...epubBuildOptions,
-    watch: {
-      onRebuild (error, result) {
-        if (error) {
-          console.error('[watch] epub build failed:', error)
-        } else {
-          console.log('[watch] epub assets rebuilt')
-          if (result.errors.length > 0) console.log('[watch] epub build errors:', result.errors)
-          if (result.warnings.length > 0) console.log('[watch] epub build warnings:', result.warnings)
-          generateDocs('epub')
-        }
-      }
+/** @type {import('esbuild').BuildOptions[]} */
+const formatters = [
+  {
+    formatter: 'epub',
+    outdir: path.resolve('../formatters/epub/dist'),
+    entryPoints: [
+      'js/entry/epub.js',
+      'css/entry/epub-elixir.css',
+      'css/entry/epub-erlang.css'
+    ]
+  },
+  {
+    formatter: 'html',
+    outdir: path.resolve('../formatters/html/dist'),
+    entryPoints: [
+      'js/entry/html.js',
+      'css/entry/html-elixir.css',
+      'css/entry/html-erlang.css'
+    ],
+    loader: {
+      '.woff2': 'file',
+      // TODO: Remove when @fontsource/* removes legacy .woff
+      '.woff': 'file'
     }
-  }).then(() => generateDocs('epub')).catch(() => process.exit(1))
-}
-
-// HTML: esbuild options
-const htmlBuildOptions = {
-  ...commonOptions,
-  outdir: htmlOutDir,
-  plugins: [htmlOnStartPlugin, loadHandlebars],
-  entryPoints: [
-    'js/entry/html.js',
-    'css/entry/html-elixir.css',
-    'css/entry/html-erlang.css'
-  ],
-  loader: {
-    '.woff2': 'file',
-    // TODO: Remove when @fontsource/* removes legacy .woff
-    '.woff': 'file'
   }
-}
+]
 
-// HTML: esbuild (conditionally configuring watch mode and rebuilding of docs)
-if (!watchMode) {
-  esbuild.build(htmlBuildOptions).catch(() => process.exit(1))
-} else {
-  esbuild.build({
-    ...htmlBuildOptions,
-    watch: {
-      onRebuild (error, result) {
-        if (error) {
-          console.error('[watch] html build failed:', error)
-        } else {
-          console.log('[watch] html assets rebuilt')
-          if (result.errors.length > 0) console.log('[watch] html build errors:', result.errors)
-          if (result.warnings.length > 0) console.log('[watch] html build warnings:', result.warnings)
-          generateDocs('html')
+Promise.all(formatters.map(async ({formatter, ...options}) => {
+  // Clean outdir.
+  await fsExtra.emptyDir(options.outdir)
+
+  await esbuild.build({
+    entryNames: watchMode ? '[name]-dev' : '[name]-[hash]',
+    bundle: true,
+    minify: !watchMode,
+    logLevel: watchMode ? 'warning' : 'info',
+    watch: watchMode,
+    ...options,
+    plugins: [{
+      name: 'ex_doc',
+      setup (build) {
+        // Pre-compile handlebars templates.
+        build.onLoad({
+          filter: /\.handlebars$/
+        }, async ({ path: filename }) => {
+          try {
+            const source = await fs.readFile(filename, 'utf-8')
+            const template = handlebars.precompile(source)
+            const contents = [
+              "import * as Handlebars from 'handlebars/runtime'",
+              "import '../helpers'",
+              `export default Handlebars.template(${template})`
+            ].join('\n')
+            return { contents }
+          } catch (error) {
+            return { errors: [{ text: error.message }] }
+          }
+        })
+
+        // Generate docs with new assets (watch mode only).
+        if (watchMode) {
+          build.onEnd(async result => {
+            if (result.errors.length) return
+            console.log(`[watch] ${formatter} assets built`)
+            await exec('mix compile --force', {cwd: '../'})
+            await exec(`mix docs --formatter ${formatter}`, {cwd: '../'})
+            console.log(`[watch] ${formatter} docs built`)
+          })
         }
       }
-    }
-  }).then(() => {
-    generateDocs('html')
-  }).catch(() => process.exit(1))
-}
-
-/**
- * Functions
- */
-
-// Docs generation (used in watch mode only)
-function generateDocs (formatter) {
-  console.log(`Building ${formatter} docs`)
-  process.chdir('../')
-  cp.execSync('mix compile --force')
-  cp.execSync(`mix docs --formatter ${formatter}`)
-  process.chdir('./assets/')
-}
+    }]
+  })
+})).catch((error) => {
+  console.error(error)
+  process.exit(1)
+})