--- a/lib/ex_doc/formatter/html/templates/head_template.eex
+++ b/lib/ex_doc/formatter/html/templates/head_template.eex
@@ -17,8 +17,6 @@
     <%= if config.canonical do %>
       <link rel="canonical" href="<%= config.canonical %>" />
     <% end %>
-    <script src="<%= asset_rev config.output, "dist/handlebars.runtime-*.js" %>"></script>
-    <script src="<%= asset_rev config.output, "dist/handlebars.templates-*.js" %>"></script>
     <script src="<%= asset_rev config.output, "dist/sidebar_items-*.js" %>"></script>
     <script src="docs_config.js"></script>
     <script async src="<%= asset_rev config.output, "dist/html-*.js" %>"></script>